+++ src/planner/rules.rs
@@ -2,254 +2,11 @@ use std::collections::HashSet;
 
 use egg::{rewrite as rw, *};
 
-use crate::array::ArrayImpl;
-use crate::catalog::ColumnRefId;
-use crate::parser::{BinaryOperator, UnaryOperator};
-use crate::types::{DataValue, PhysicalDataTypeKind};
-
-define_language! {
-    pub enum Plan {
-        Constant(DataValue),
-        Type(PhysicalDataTypeKind),
-        ColumnRef(ColumnRefId),
-        // "*" = Wildcard,
-
-        // binary operations
-        "+" = Add([Id; 2]),
-        "-" = Sub([Id; 2]),
-        "*" = Mul([Id; 2]),
-        "/" = Div([Id; 2]),
-        "%" = Mod([Id; 2]),
-        "||" = StringConcat([Id; 2]),
-        ">" = Gt([Id; 2]),
-        "<" = Lt([Id; 2]),
-        ">=" = GtEq([Id; 2]),
-        "<=" = LtEq([Id; 2]),
-        "=" = Eq([Id; 2]),
-        "<>" = NotEq([Id; 2]),
-        "and" = And([Id; 2]),
-        "or" = Or([Id; 2]),
-        "xor" = Xor([Id; 2]),
-        "like" = Like([Id; 2]),
-
-        // unary operations
-        "-" = Neg(Id),
-        "not" = Not(Id),
-        "isnull" = IsNull(Id),
-
-        // aggregate
-        "max" = Max(Id),
-        "min" = Min(Id),
-        "sum" = Sum(Id),
-        "avg" = Avg(Id),
-        "count" = Count(Id),
-        "rowcount" = RowCount(Id),
-        "first" = First(Id),
-        "last" = Last(Id),
-
-        "cast" = TypeCast([Id; 2]),
-        "as" = Alias([Id; 2]),
-        "fn" = Function(Box<[Id]>),
-
-        "select" = Select([Id; 9]),             // (select
-                                                //      distinct=[expr..]
-                                                //      select_list=[expr..]
-                                                //      from=[table..]
-                                                //      where=expr
-                                                //      groupby=[expr..]
-                                                //      having=expr
-                                                //      orderby=[expr..]
-                                                //      limit=expr
-                                                //      offset=expr
-                                                // )
-
-        "scan" = Scan([Id; 2]),                 // (scan table [column..])
-        "values" = Values(Box<[Id]>),           // (values tuple..)
-        "proj" = Proj([Id; 2]),                 // (proj [expr..] child)
-        "filter" = Filter([Id; 2]),             // (filter expr child)
-        "order" = Order([Id; 2]),               // (order [order_key..] child)
-            "order_key" = OrderKey([Id; 2]),        // (order_key expr asc/desc)
-                "asc" = Asc,
-                "desc" = Desc,
-        "limit" = Limit([Id; 3]),               // (limit offset limit child)
-        "topn" = TopN([Id; 4]),                 // (topn offset limit [order_key..] child)
-        "join" = Join([Id; 4]),                 // (join join_type expr left right)
-        "hashjoin" = HashJoin([Id; 5]),         // (hashjoin join_type [left_expr..] [right_expr..] left right)
-            "inner" = Inner,
-            "left_outer" = LeftOuter,
-            "right_outer" = RightOuter,
-            "full_outer" = FullOuter,
-            "cross" = Cross,
-        "agg" = Agg([Id; 3]),                   // (agg aggs=[expr..] group_keys=[expr..] child)
-        "create" = Create([Id; 2]),             // (create table [column_desc..])
-        "drop" = Drop(Id),                      // (drop table)
-        "insert" = Insert([Id; 3]),             // (insert table [column..] child)
-        "delete" = Delete([Id; 2]),             // (delete table child/true)
-        "copy_from" = CopyFrom(Id),             // (copy_from dest)
-        "copy_to" = CopyTo([Id; 2]),            // (copy_to dest child)
-        "explain" = Explain(Id),                // (explain child)
-
-        "tuple" = Tuple(Box<[Id]>),             // (tuple expr..)
-        "list" = List(Box<[Id]>),               // (list ...)
-
-        Symbol(Symbol),
-    }
-}
-
-impl Plan {
-    const fn binary_op(&self) -> Option<(BinaryOperator, Id, Id)> {
-        use BinaryOperator as Op;
-        Some(match self {
-            &Plan::Add([a, b]) => (Op::Plus, a, b),
-            &Plan::Sub([a, b]) => (Op::Minus, a, b),
-            &Plan::Mul([a, b]) => (Op::Multiply, a, b),
-            &Plan::Div([a, b]) => (Op::Divide, a, b),
-            &Plan::Mod([a, b]) => (Op::Modulo, a, b),
-            &Plan::StringConcat([a, b]) => (Op::StringConcat, a, b),
-            &Plan::Gt([a, b]) => (Op::Gt, a, b),
-            &Plan::Lt([a, b]) => (Op::Lt, a, b),
-            &Plan::GtEq([a, b]) => (Op::GtEq, a, b),
-            &Plan::LtEq([a, b]) => (Op::LtEq, a, b),
-            &Plan::Eq([a, b]) => (Op::Eq, a, b),
-            &Plan::NotEq([a, b]) => (Op::NotEq, a, b),
-            &Plan::And([a, b]) => (Op::And, a, b),
-            &Plan::Or([a, b]) => (Op::Or, a, b),
-            &Plan::Xor([a, b]) => (Op::Xor, a, b),
-            &Plan::Like([a, b]) => (Op::Like, a, b),
-            _ => return None,
-        })
-    }
-
-    const fn unary_op(&self) -> Option<(UnaryOperator, Id)> {
-        use UnaryOperator as Op;
-        Some(match self {
-            &Plan::Neg(a) => (Op::Minus, a),
-            &Plan::Not(a) => (Op::Not, a),
-            _ => return None,
-        })
-    }
-}
-
-#[derive(Default)]
-pub struct PlanAnalysis;
-
-#[derive(Debug)]
-pub struct Data {
-    /// Some if the expression is a constant.
-    val: Option<DataValue>,
-    /// All columns involved in the node.
-    columns: ColumnSet,
-    /// The schema for plan node: a list of expressions.
-    schema: Option<Vec<Id>>,
-}
-
-type ColumnSet = HashSet<ColumnRefId>;
-
-impl Analysis<Plan> for PlanAnalysis {
-    type Data = Data;
-
-    fn merge(&mut self, to: &mut Self::Data, from: Self::Data) -> DidMerge {
-        let merge_val = egg::merge_max(&mut to.val, from.val);
-        let merge_col = merge_columns(&mut to.columns, from.columns);
-        let merge_schema = egg::merge_max(&mut to.schema, from.schema);
-        merge_val | merge_col | merge_schema
-    }
-
-    fn make(egraph: &EGraph, enode: &Plan) -> Self::Data {
-        Data {
-            val: eval(egraph, enode),
-            columns: analyze_columns(egraph, enode),
-            schema: analyze_schema(egraph, enode),
-        }
-    }
-
-    fn modify(egraph: &mut EGraph, id: Id) {
-        // add a new constant node
-        if let Some(val) = &egraph[id].data.val {
-            let added = egraph.add(Plan::Constant(val.clone()));
-            egraph.union(id, added);
-        }
-    }
-}
-
-/// Evaluate constant.
-fn eval(egraph: &EGraph, enode: &Plan) -> Option<DataValue> {
-    use Plan::*;
-    let x = |i: Id| egraph[i].data.val.as_ref();
-    if let Constant(v) = enode {
-        Some(v.clone())
-    } else if let Some((op, a, b)) = enode.binary_op() {
-        let array_a = ArrayImpl::from(x(a)?);
-        let array_b = ArrayImpl::from(x(b)?);
-        Some(array_a.binary_op(&op, &array_b).get(0))
-    } else if let Some((op, a)) = enode.unary_op() {
-        let array_a = ArrayImpl::from(x(a)?);
-        Some(array_a.unary_op(&op).get(0))
-    } else if let &IsNull(a) = enode {
-        Some(DataValue::Bool(x(a)?.is_null()))
-    } else if let &TypeCast(_) = enode {
-        // TODO: evaluate type cast
-        None
-    } else if let &Max(a) | &Min(a) | &Avg(a) | &First(a) | &Last(a) = enode {
-        x(a).cloned()
-    } else {
-        None
-    }
-}
-
-/// Returns all columns involved in the node.
-fn analyze_columns(egraph: &EGraph, enode: &Plan) -> ColumnSet {
-    if let Plan::ColumnRef(col) = enode {
-        return [*col].into_iter().collect();
-    }
-    // merge the set from all children
-    (enode.children().iter())
-        .flat_map(|id| egraph[*id].data.columns.iter().cloned())
-        .collect()
-}
-
-/// Merge 2 column set.
-fn merge_columns(to: &mut ColumnSet, from: ColumnSet) -> DidMerge {
-    let len = to.len();
-    if from.is_subset(to) {
-        *to = from;
-        DidMerge(to.len() != len, false)
-    } else {
-        DidMerge(false, true)
-    }
-}
-
-/// Returns the output expressions for plan node.
-fn analyze_schema(egraph: &EGraph, enode: &Plan) -> Option<Vec<Id>> {
-    use Plan::*;
-    let x = |i: Id| egraph[i].data.schema.clone().unwrap();
-    let concat = |v1: Vec<Id>, v2: Vec<Id>| v1.into_iter().chain(v2.into_iter()).collect();
-    Some(match enode {
-        // equal to child
-        Filter([_, c]) | Order([_, c]) | Limit([_, _, c]) | TopN([_, _, _, c]) => x(*c),
-
-        // concat 2 children
-        Join([_, _, l, r]) | HashJoin([_, _, _, l, r]) => concat(x(*l), x(*r)),
-
-        // list is the source for the following nodes
-        List(ids) => ids.to_vec(),
-
-        // plans that change schema
-        Scan([_, columns]) => x(*columns),
-        Values(_) => todo!("add schema for values plan"),
-        Proj([exprs, _]) => x(*exprs),
-        Agg([exprs, group_keys, _]) => concat(x(*exprs), x(*group_keys)),
-
-        // not plan node
-        _ => return None,
-    })
-}
-
-pub type EGraph = egg::EGraph<Plan, PlanAnalysis>;
-pub type Rewrite = egg::Rewrite<Plan, PlanAnalysis>;
+use super::{ColumnSet, EGraph, Rewrite};
+use crate::types::DataValue;
 
 #[rustfmt::skip]
-pub fn rules() -> Vec<Rewrite> { vec![
+fn expr_rules() -> Vec<Rewrite> { vec![
     rw!("add-zero";  "(+ ?a 0)" => "?a"),
     rw!("add-comm";  "(+ ?a ?b)" => "(+ ?b ?a)"),
     rw!("add-assoc"; "(+ ?a (+ ?b ?c))" => "(+ (+ ?a ?b) ?c)"),
@@ -322,7 +79,10 @@ pub fn rules() -> Vec<Rewrite> { vec![
     rw!("or-assoc";  "(or ?a (or ?b ?c))" => "(or (or ?a ?b) ?c)"),
 
     rw!("avg";       "(avg ?a)" => "(/ (sum ?a) (count ?a))"),
+]}
 
+#[rustfmt::skip]
+fn plan_rules() -> Vec<Rewrite> { vec![
     rw!("limit-order=topn";
         "(limit ?offset ?limit (order ?keys ?child))" =>
         "(topn ?offset ?limit ?keys ?child)"
@@ -426,25 +186,25 @@ fn columns_is(
 
 egg::test_fn! {
     and_eq_const,
-    rules(),
+    expr_rules(),
     "(and (= a 1) (= a b))" => "(and (= a 1) (= b 1))",
 }
 
 egg::test_fn! {
     constant_folding,
-    rules(),
+    expr_rules(),
     "(* (- (+ 1 2) 4) (/ 6 2))" => "-3",
 }
 
 egg::test_fn! {
     constant_moving,
-    rules(),
+    expr_rules(),
     "(> (+ 100 a) 300)" => "(> a 200)",
 }
 
 egg::test_fn! {
     predicate_pushdown,
-    rules(),
+    plan_rules(),
     // SELECT s.name, e.cid
     // FROM student AS s, enrolled AS e
     // WHERE s.sid = e.sid AND e.grade = 'A'
@@ -477,7 +237,7 @@ egg::test_fn! {
 
 egg::test_fn! {
     join_reorder,
-    rules(),
+    plan_rules(),
     // SELECT * FROM t1, t2, t3
     // WHERE t1.id = t2.id AND t3.id = t2.id
     "
@@ -511,7 +271,7 @@ egg::test_fn! {
 
 egg::test_fn! {
     hash_join,
-    rules(),
+    plan_rules(),
     // SELECT * FROM t1, t2
     // WHERE t1.id = t2.id AND t1.age > 2
     "
@@ -538,7 +298,7 @@ egg::test_fn! {
 
 egg::test_fn! {
     agg,
-    rules(),
+    plan_rules(),
     // SELECT sum(a + b) + count(a) FROM t
     // GROUP BY a;
     "
